Split toplevel parsing
